+++ Realization/frontend/czechidm-example/src/content/dashboards/ExampleDashboard.js
@@ -18,24 +18,28 @@ export default class ExampleDashboard extends Basic.AbstractContent {
 
   render() {
     return (
-      <Basic.Panel>
-        <Basic.PanelHeader text={ this.i18n('header') }/>
-        <Basic.PanelBody>
-          <Basic.Panel className="panel-warning no-margin">
-            <Basic.PanelHeader>
-              <Basic.Row>
-                <Basic.Col lg={ 3 }>
-                  <Basic.Icon type="fa" icon="dashboard" className="fa-5x"/>
-                </Basic.Col>
-                <Basic.Col lg={ 9 }>
-                  <div><strong>{ this.i18n('title') }</strong></div>
-                  <div>{ this.i18n('text') }</div>
-                </Basic.Col>
-              </Basic.Row>
-            </Basic.PanelHeader>
-          </Basic.Panel>
-        </Basic.PanelBody>
-      </Basic.Panel>
+      <div>
+        <Basic.ContentHeader
+          icon="gift"
+          text={ this.i18n('header') }/>
+        <Basic.Panel>
+          <Basic.PanelBody>
+            <Basic.Panel className="panel-warning no-margin">
+              <Basic.PanelHeader>
+                <Basic.Row>
+                  <Basic.Col lg={ 1 }>
+                    <Basic.Icon type="fa" icon="dashboard" className="fa-5x"/>
+                  </Basic.Col>
+                  <Basic.Col lg={ 11 }>
+                    <div><strong>{ this.i18n('title') }</strong></div>
+                    <div>{ this.i18n('text') }</div>
+                  </Basic.Col>
+                </Basic.Row>
+              </Basic.PanelHeader>
+            </Basic.Panel>
+          </Basic.PanelBody>
+        </Basic.Panel>
+      </div>
     );
   }
 }